requirements.txt
pydantic==1.9.0
fastapi==0.79.1
uvicorn==0.17.6
mysql-connector-python==8.0.30
pytest==7.1.2